Adding zoom to level.

+++ Assets/Code/Ships/Camino.cs
@@ -4,10 +4,10 @@ namespace RoundedDefence
 {
 	public class Camino
 	{
-		public byte lvl=0;
-		public byte p=0;
+		public int lvl=0;
+		public int p=0;
 		public int t=0;
-		public Camino(byte l,byte pp,int tt){
+		public Camino(int l,int pp,int tt){
 			lvl=l;
 			p=pp;
 			t=tt;
